put configSetting in its own file and defined join-msgs-enabled server setting
